refactor: Add App component
